#### [REM] core: remove `BaseModel/Connection.__nonzero__`

`__nonzero__` is only used by Python 2. Time to remove them.

#### [IMP] base: improve performance of res.partner.child_ids field

Reading res.partner.child_ids is slower than necessary and slows down `_handle_first_contact_creation` (useful in the populate tool) and `address_get`. In fact, the domain `[('active', '=', True)]` on the field already filters out inactive records during the `One2many.read`, but the ORM does the same job (in Python) every time the field is accessed (see `One2many.convert_to_record`).

Avoid this extra filter by adding `context={'active_test': False}` in the field. Also, reorder the condition in `_handle_first_contact_creation` to avoid reading child_ids most of the time (much more expensive than the other conditions).

#### [IMP] core: improve performance of `BaseModel.__getitem__`

Improve the performance of `BaseModel.__getitem__` (heavily used by the ORM) and avoid an useless lookups on self.env.registry. The owner attribute is not used in `Field.__get__`. The Python documentation mentions that the 'owner' (objtype) attribute should have a default value of None, which allows us to omit it entirely when calling from
`BaseModel.__getitem__` (see https://docs.python.org/3.9/howto/descriptor.html#descriptor-protocol).

Before `self['id']` took taking 183 ns ± 0.142 ns and now it's taking 121 ns ± 0.396 ns. Note that `self.id` took 74.1 ns ± 0.0983 ns, then the extra cost of `BaseModel.__getitem__` is reduced from 109 ns to 47 ns.

#### [IMP] core: improve performance of `BaseModel.filtered`

Improve the performance of BaseModel.filtered() when the func attribute is a string. This is especially important for _RelationalMulti.convert_to_record(), which uses filtered('active') every time an X2Many is accessed. Avoid the extra work done by BaseModel.mapped() for simple filtering, which is called on each single record.

Performance Test (cache full):
```
records_239 = self.env['ir.ui.view'].with_context(active_test=False).search([])
records_10 = records_239[:10]
record = records_239[0]
```

Before:
```
record.filtered('active'):        2.29 µs ±  42.2 ns
records_10.filtered('active'):   18    µs ±  45.3 ns
records_239.filtered('active'): 402    µs ± 610   ns
```
Now:
```
record.filtered('active'):        1.37 µs ±   1.2 ns
records_10.filtered('active'):    8.74 µs ±  41.4 ns
records_239.filtered('active'): 193    µs ± 415   ns
```
#### [IMP] core: slightly improve performance of the Cache fields

Slightly improve the performance of the field cache to improve the performance of Odoo in general:

- Add `__slots__` to the Cache class.
- Change all conditions `model.pool.field_depends_context[field]` to `field in model.pool.field_depends_context` to avoid extra calls to `get` done in `Collector.__getitem__` and because the Collector doesn't hold empty tuples.

Before `self.login` took 445 ns ± 0.868 ns and with the improvement, it takes 405 ns ± 0.437 ns.

#### [REF] core: refactor env attributes

- Remove the Environment.all attribute, it was a duplicate of Environment.transaction (which is more verbose).
- Remove Environment.args, not used anymore, and don't bring it back for its original purpose, because a quick test shows that it was mostly useless.
- Remove the useless assert on context in `Environment.__new__`, it will be checked when transforming to `frozendict`.




Futher improvements can be done for `Field.__get__`: https://github.com/odoo/odoo/pull/168646/files