@robodoo 2024-06-24
Merge method set to rebase and fast-forward.